requirements.txt
# requirements.txt for Smart Resume Screener
pdfplumber
python-docx
nltk
spacy
scikit-learn
pandas
